+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Improvement: Add hidden `serverOptions` and `bodyParser` configurations to `http` configuration file;
 - Improvement: Add `.gitkeep` files to each `api` folder;
 - Improvement: Extend `api` folder with other directories which loads with Sails by default;
+- Improvement: Implement checking for `generator-sails-rest-api` updates in CLI;
 - Improvement: Optimizing and improvement yeoman `index.js` file structure;
 - Improvement: Automatically creates `.yo-rc.json` file where generator's configuration is storing;
 - Fix: Add `skip-install` to yeoman as it should be declared (before it was wrong);